clib: use internaly bool type and true/false values

+++ bld/clib/conio/c/getchos2.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2019 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2025 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -64,7 +64,7 @@ _WCRTLINK int getch( void )
 #ifdef DEFAULT_WINDOWING
     if( _WindowsGetch != NULL ) {   // Default windowing
         LPWDATA     res;
-        res = _WindowsIsWindowedHandle( (int)STDIN_FILENO );
+        res = _WindowsIsWindowedHandle( STDIN_FILENO );
         return( _WindowsGetch( res ) );
     }
 #endif
@@ -80,7 +80,8 @@ _WCRTLINK int getch( void )
     rc = KbdCharIn( &info, 0, 0 );
     if( rc == ERROR_KBD_DETACHED )
         return( EOF );
-    if( info.chChar == 0 || info.chChar == 0xe0 ) {
+    if( info.chChar == 0
+      || info.chChar == 0xe0 ) {
         _RWD_cbyte2 = info.chScan;
     }
     return( info.chChar );
